Inheritance correction

+++ experiment_control/scan_generator.py
@@ -59,15 +59,15 @@ class ScanGenerator:
 
 
 
-    def run_repeats(self, N,lims=None, save = 'npy'):
+    def run_repeats(self, N,lims=None, save = 'npy', dtype=np.uint16):
         for i in tqdm(range(N)):
 
             timestamp = time.strftime("-%Y%m%d-%H%M%S")
             # image = Image.fromarray(self.detector.photo_capture().astype(np.uint8))
             if lims==None:
-                arr_new =  self.detector.photo_capture().astype(np.uint8)
+                arr_new =  self.detector.photo_capture().astype(dtype)
             else:
-                arr_new =  self.detector.photo_capture().astype(np.uint8)[lims[0]:lims[1],lims[2]:lims[3]]
+                arr_new =  self.detector.photo_capture().astype(dtype)[lims[0]:lims[1],lims[2]:lims[3]]
 
             if save=='npy':
                 filename = os.path.join(self.save_directory, str(i)+timestamp +  ".npy")
@@ -81,7 +81,7 @@ class ScanGenerator:
                 im.save(filename, dpi=(300,300))
                 print('Saved', filename)
 
-    def run_scan(self, repeat=1,lims=None):
+    def run_scan(self, repeat=1,lims=None, dtype = np.uint16):
 
         self.make_variable_space()
         flat_lists = [[item for sublist in self.stage_list[key] for item in sublist] for key in list(self.stage_list.keys())]
@@ -106,9 +106,9 @@ class ScanGenerator:
             # take photo
             # image = Image.fromarray(self.detector.photo_capture_repeat(repeat).astype(np.uint8))
             if lims==None:
-                arr_new =  self.detector.photo_capture_repeat(repeat).astype(np.uint8)
+                arr_new =  self.detector.photo_capture_repeat(repeat).astype(dtype)
             else:
-                arr_new =  self.detector.photo_capture_repeat(repeat).astype(np.uint8)[lims[0]:lims[1],lims[2]:lims[3]]
+                arr_new =  self.detector.photo_capture_repeat(repeat).astype(dtype)[lims[0]:lims[1],lims[2]:lims[3]]
             filename = os.path.join(self.save_directory, str(i)+timestamp + ".npy")
             #image.save(filename)
             np.save(filename, arr_new)
